+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Fixed assertions on empty functions [GH-50]
   - Fixed New-Fixture so it creates proper syntax in tests [GH-49]
   - Fixed assertions on Object arrays [GH-61]
+  - Fixed issue where curly brace misalignment would cause issues [GH-90]
 
 ## 2.0.3 (Apr 16, 2013)